ml-inference-platform/inference-service/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
pydantic==2.4.2
pydantic-settings==2.0.3
huggingface-hub==0.19.4
joblib==1.3.2
scikit-learn==1.3.2
